Add client creation functionality to seed scripts
Added the 'createClient' function in the 'seed-utils.ts' script and integrated it into the 'seed.ts' script. This new function generates fake client data using '@faker-js/faker' and inserts this information into the 'clients' database table. The function is invoked in 'seed.ts' only if the current user has an 'Admin' role.