CICD (#27)
* Add anonymous key to GitHub Actions workflow

The commit introduces a new environment variable in the ci.yml file for GitHub Actions, which is the PUBLIC_SUPABASE_ANON_KEY. This allows the anonymous use of the Supabase API, which is essential for certain tasks in the continuous integration workflow.

* Introduce PUBLIC_SUPABASE_URL to workflow

The commit includes a new environment variable - PUBLIC_SUPABASE_URL in the GitHub Actions workflow. This adjusts settings within ci.yml for calls to the Supabase API.

* Add SUPABASE_SERVICE_ROLE_KEY to GitHub Actions workflow

A new environment variable, SUPABASE_SERVICE_ROLE_KEY, has been added to the GitHub Actions workflow inside the ci.yml file. This change allows for more robust interactions with the Supabase API.