Refactor GitHub workflows and split into separate files
The single GitHub workflow file `ci.yml` was deleted and the workflow was split into two new files: `pr-opened-or-commits-pushed.yml` and `tag-created.yml`. This is to separate concerns and make it easier to manage and understand the operations happening on different triggers such as PR opening/commit pushing and tag creation.