Update GitHub workflow triggers for tags
The GitHub workflow file `.github/workflows/tag-created.yml` has been updated to trigger every time a new tag is created. Previously, it was triggering with every push to the main branch. This update increases the versatility and control of our workflow processes by allowing actions to be triggered with specific tags.