CICD (#31)
* Add .vercel to .gitignore

This commit introduces .vercel to the .gitignore file. It's an important modification that will prevent unwanted tracking and sharing of .vercel files whenever changes are pushed to the repository.

* Modify permissions in GitHub workflows, update "release" action

This update introduces read and write permissions for specific actions within the release job in GitHub workflows. Further, the "release" action in package.json has been modified to properly version changesets instead of directly publishing.