CICD (#32)
* Add .vercel to .gitignore

This commit introduces .vercel to the .gitignore file. It's an important modification that will prevent unwanted tracking and sharing of .vercel files whenever changes are pushed to the repository.

* Modify permissions in GitHub workflows, update "release" action

This update introduces read and write permissions for specific actions within the release job in GitHub workflows. Further, the "release" action in package.json has been modified to properly version changesets instead of directly publishing.

* Disable multiple steps in release workflow

This commit comments out several steps in the release workflow including starting local Supabase, installing Playwright Browsers, running unit and integration tests, archiving test reports, and several deployment steps. The changes appear to be a part of a bigger restructuring of the workflow process.