CICD (#35)
* Add .vercel to .gitignore

This commit introduces .vercel to the .gitignore file. It's an important modification that will prevent unwanted tracking and sharing of .vercel files whenever changes are pushed to the repository.

* Modify permissions in GitHub workflows, update "release" action

This update introduces read and write permissions for specific actions within the release job in GitHub workflows. Further, the "release" action in package.json has been modified to properly version changesets instead of directly publishing.

* Disable multiple steps in release workflow

This commit comments out several steps in the release workflow including starting local Supabase, installing Playwright Browsers, running unit and integration tests, archiving test reports, and several deployment steps. The changes appear to be a part of a bigger restructuring of the workflow process.

* Update GitHub token in release workflow

This commit updates the GitHub token used in the "Create Release Pull Request" step of the release workflow. The change switches from the custom `GH_TOKEN` to the standard `GITHUB_TOKEN`, ensuring compatibility across various GitHub actions.

* Remove tag-created workflow and modify release workflow

The tag-created workflow file is deleted and the release workflow file is updated to revert commented out sections, enabling functions such as starting local Supabase, installing Playwright Browsers, and deploying to Vercel. Changeset configuration and script in package.json are also altered as part of this update.

* Enable tag creation and deployment to Vercel

Changes include the addition of automatic tag creation and deployment to Vercel capabilities. This comes as part of the broader workflow updates, such as the removal of the tag-created workflow and modifications to the release workflow, including re-enabled commented out sections like local Supabase initiation and Playwright Browsers installation. Updates also impact Changeset configuration and scripts in package.json.