CICD (#48)
* Refactor GitHub workflows for app release and deployment

This commit removes the redundant 'tag-created.yml' workflow and merges its functionalities into 'release.yml', optimizing the workflow for deploying new versions of the application. As such, the up-to-date 'release.yml' now contains the necessary commands to seamlessly and efficiently update the deployed app.

* Update pnpm commands in GitHub workflows

This commit modifies the 'release.yml' GitHub workflow file and updates the 'version' and 'publish' commands to use 'pnpm changeset'. This change is intended to enhance and streamline the app release and deployment process. Furthermore, a new command 'createGithubReleases' is also added to facilitate automatic creation of GitHub releases.