CICD (#55)
* Refactor GitHub workflows for app release and deployment

This commit removes the redundant 'tag-created.yml' workflow and merges its functionalities into 'release.yml', optimizing the workflow for deploying new versions of the application. As such, the up-to-date 'release.yml' now contains the necessary commands to seamlessly and efficiently update the deployed app.

* Update pnpm commands in GitHub workflows

This commit modifies the 'release.yml' GitHub workflow file and updates the 'version' and 'publish' commands to use 'pnpm changeset'. This change is intended to enhance and streamline the app release and deployment process. Furthermore, a new command 'createGithubReleases' is also added to facilitate automatic creation of GitHub releases.

* Modify pnpm commands in release.yml workflow

This commit updates the 'version' and 'publish' pnpm commands within the 'release.yml' workflow. It is intended to streamline the release and deployment process. The 'createGithubReleases' command has also been included for auto-generation of GitHub releases.

* Simplify commands in GitHub release workflow

The commands 'version' and 'publish' in the 'release.yml' GitHub workflow were amended to 'pnpm ci:version' and 'pnpm ci:publish' respectively. Doing so aims to make the release and deployment procedure more efficient. The feature to automatically generate GitHub releases, 'createGithubReleases', is also now included.

* Remove redundant commands from package.json

The commands 'version' and 'release' have been removed from the scripts section in package.json. The 'ci:version' command has been updated to use 'changeset version'. This simplification streamlines the script commands and might improve the execution of continuous integration workflows.

* Added support for private packages in changesets

In the .changeset/config.json file, a new configuration for "privatePackages" with options for "version" and "tag" were added. Additionally, the script "ci:publish" in package.json was updated to use "changeset publish" instead of "pnpm publish -r", potentially improving the continuous integration process.