Update package json to show git repository (#63)
* Add repository information to package.json

Repository details were added to the package.json file for the "prive" project. Specifically, the type of the repository is defined as "git", and the URL to access the git repository was provided. This change helps to identify and access the source code repository directly from package.json.

* Add repository information to package.json

Added repository details to the `package.json` file in the "prive" project. The repository type is specified as 'git' with provided URL. This ensures easier access to the source code repository directly from the package.json.

* Correct indentation in package.json

This commit makes sure the indentation is consistent in the package.json file. The "engines" section has been re-indented for better readability and maintenance. The file makes use of tabs for indentation.

* Correct indentation in package.json

This commit makes sure the indentation is consistent in the package.json file. The "engines" section has been re-indented for better readability and maintenance. The file makes use of tabs for indentation.