Feature/deploy if tag is created (#69)
* Update release workflow with conditional deploy step

The release workflow within the GitHub repository was updated. Specifically, a condition to check if any changesets were published was included. In case of any publication, a deployment is triggered immediately; adding another layer of automation to the workflow.

* Add changeset presence check to CI workflow

The CI workflow file (.github/workflows/ci.yml) has been updated to include a check for the presence of a changeset. If a changeset is found, indicating that changes have been published, an automatic deployment will be triggered. This adds an extra layer of automation to our release process.

* add changesets

* Update CI workflow to use pnpm for changeset status check

The updated .github/workflows/ci.yml file now uses pnpm to check changeset status in comparison to the main branch within continuous integration workflow. This brings the benefit of consistency especially if pnpm is being extensively used throughout the project.

* Update CI workflow to fetch all Git history

In the .github/workflows/ci.yml file, an additional parameter has been added to the action checkout step which fetches the entirety of the Git history. This allows actions such as Changesets to generate accurate changelogs, as all the commits are now present.

* Disable changeset check in CI workflow

The changeset check in the Continuous Integration workflow configured in .github/workflows/ci.yml is commented out. This step, which used to check for changesets since the main branch, has been disabled for the time being. Future updates will determine if it needs to be re-enabled.

* Remove changeset check from CI workflow

The change involves removal of the changeset presence check from the Continuous Integration (CI) workflow. This removal is due to the commenting out of the respective lines in .github/workflows/ci.yml. The need for this check may be evaluated in future based on the project's requirements.