Hide "Register" button on production
To maintain the user registration process strictly on the development environment, a conditional statement was added in the Navigation component file to display the "Register" button only when running in development mode. Moreover, the registration route server layout was also modified not to load in production mode, it will instead redirect to the homepage.